Se culminó con la creacion de los archivos Repository de Historial Sesiones
